Deduplicate missing entitlement keychain fixture
